refactor web to align up codebase
