+++ apps/web/package.json
@@ -14,6 +14,7 @@
     "@types/react-syntax-highlighter": "^15.5.6",
     "@vercel/analytics": "^0.1.11",
     "@vercel/examples-ui": "^1.0.5",
+    "@airthium/extract-json-from-string": "1.0.8",
     "clsx": "^1.2.1",
     "next": "latest",
     "react": "latest",
