+++ apps/web/src/core/components/Chat.tsx
