+++ apps/web/src/core/components/hooks/useAgent.ts
@@ -1,0 +1,7 @@
+import { useWindowAI } from "./useWindowAI"
+
+export const useAgent = () => {
+  const {} = useWindowAI([])
+
+  return {}
+}
