+++ apps/web/src/core/components/hooks/useWindowAI.ts
