+++ apps/web/src/core/components/icons/Discord.tsx
