+++ apps/web/src/pages/playground.tsx
@@ -1,9 +1,9 @@
 import { Layout } from "@vercel/examples-ui"
-import { InputMessage } from "~components/InputMessage"
-import { useWindowAI } from "~components/hooks/useWindowAI"
+import { InputMessage } from "~core/components/InputMessage"
+import { useWindowAI } from "~core/components/hooks/useWindowAI"
 import clsx from "clsx"
 
-const Agent = ({ name = "Agent Name", purpose = "Purpose" }) => {
+const AgentMonitor = ({ name = "Agent Name", purpose = "Purpose" }) => {
   return (
     <div className="w-full sm:w-1/2 md:w-1/3 lg:w-1/4 p-2">
       <div
@@ -55,48 +55,7 @@ function PlaygroundPage() {
       </div>
 
       <div className="flex flex-wrap w-full">
-        <Agent
-          name="Manager"
-          purpose="Facilitate Primary Goal and Hiring new Agent"
-        />
-        <Agent name="Agent 1" purpose="Pass Butter" />
-        <Agent
-          name="Manager"
-          purpose="Facilitate Primary Goal and Hiring new Agent"
-        />
-        <Agent
-          name="Manager"
-          purpose="Facilitate Primary Goal and Hiring new Agent"
-        />
-        <Agent
-          name="Manager"
-          purpose="Facilitate Primary Goal and Hiring new Agent"
-        />
-        <Agent
-          name="Manager"
-          purpose="Facilitate Primary Goal and Hiring new Agent"
-        />
-        <Agent
-          name="Manager"
-          purpose="Facilitate Primary Goal and Hiring new Agent"
-        />
-        <Agent
-          name="Manager"
-          purpose="Facilitate Primary Goal and Hiring new Agent"
-        />
-        <Agent
-          name="Manager"
-          purpose="Facilitate Primary Goal and Hiring new Agent"
-        />
-        <Agent
-          name="Manager"
-          purpose="Facilitate Primary Goal and Hiring new Agent"
-        />
-        <Agent
-          name="Manager"
-          purpose="Facilitate Primary Goal and Hiring new Agent"
-        />
-        <Agent
+        <AgentMonitor
           name="Manager"
           purpose="Facilitate Primary Goal and Hiring new Agent"
         />
