+++ apps/web/tailwind.config.js
@@ -1,8 +1,7 @@
 module.exports = {
   presets: [require("@vercel/examples-ui/tailwind")],
   content: [
-    "./pages/**/*.{js,ts,jsx,tsx}",
-    "./components/**/*.{js,ts,jsx,tsx}",
+    "./src/**/*.{js,ts,jsx,tsx}",
     "./node_modules/@vercel/examples-ui/**/*.js"
   ],
   plugins: [require("windy-radix-palette")]
